other: small refactorization, moved exceptions to seperate module, additional test for mulit source pipeline
